+++ tests/test_plugins.py
@@ -17,7 +17,14 @@ def test_noqa_fixture_disabled(absolute_path):
     """End-to-End test to check that all plugins are enabled."""
     filename = absolute_path('fixtures', 'external_plugins.py')
     process = subprocess.Popen(
-        ['flake8', '--disable-noqa', filename],
+        [
+            'flake8',
+            '--disable-noqa',
+            '--isolated',
+            '--enable-extensions',
+            'G',
+            filename,
+        ],
         stdout=subprocess.PIPE,
         stderr=subprocess.PIPE,
         universal_newlines=True,
@@ -44,4 +51,4 @@ def test_noqa_fixture_disabled(absolute_path):
     assert output.count('S101') > 0  # flake8-bandit
     assert output.count('T001') > 0  # flake8-print
     assert output.count('T100') > 0  # flake8-debugger
-    assert output.count('T800') > 0  # flake8-type-annotations
+    assert output.count('TAE002') > 0  # flake8-annotations-complexity